PDM_Manager_Software/requirements.txt
pyserial>=3.5
customtkinter>=5.2.0
pillow>=10.0.0
pyinstaller>=5.13.0
requests>=2.31.0